pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "torch-split"
version = "0.1.0"
description = "PyTorch model analysis and splitting tool"
requires-python = ">=3.9"

dependencies = [
    "torch>=2.8.0",
    "frozendict>=2.4.6",
    "graphviz>=0.21.0",
    "opentelemetry-api>=1.37.0",
    "opentelemetry-sdk>=1.37.0",
    "opentelemetry-exporter-otlp-proto-http>=1.37.0",
    "pydantic>=2.12.0",
    "msgpack>=1.1.2",
    "rich>=14.2.0",
    "unionfind>=0.2.1",
]

[project.optional-dependencies]
typing = ["types-frozendict"]

[tool.ruff]
line-length = 120
